Changes Updated for Login and Register Screens
