+++ Login-Register-Screens-UI/app/src/main/AndroidManifest.xml
