+++ Login-Register-Screens-UI/app/src/main/java/com/example/login_register_screen_ui/Login.java
@@ -1,0 +1,15 @@
+package com.example.login_register_screen_ui;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import android.os.Bundle;
+import android.view.WindowManager;
+
+public class Login extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_login);
+    }
+}
