+++ Login-Register-Screens-UI/app/src/main/res/drawable/register_img.png
