+++ Login-Register-Screens-UI/app/src/main/res/values-v21/themes.xml
