--- Register-And-Login-Screen/app/src/main/java/com/example/login_register_screen_ui/Login.java
@@ -1,17 +1,0 @@
-package com.example.login_register_screen_ui;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.os.Bundle;
-import android.view.WindowManager;
-
-public class Login extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        getWindow().setFlags(WindowManager.LayoutParams.FLAG_FULLSCREEN,WindowManager.LayoutParams.FLAG_FULLSCREEN);
-        setContentView(R.layout.welcome_screen);
-        setContentView(R.layout.activity_login);
-    }
-}
