--- Register-And-Login-Screen/app/src/main/res/layout/activity_login.xml
@@ -1,97 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical"
-    android:gravity="center_horizontal"
-    tools:context=".MainActivity">
-
-    <ImageView
-        android:layout_width="300dp"
-        android:layout_height="250dp"
-        android:src="@drawable/img_2"
-        android:id="@+id/image"
-        android:layout_margin="50dp"
-        android:transitionName="welcome_image"/>
-
-    <TextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="Login"
-        android:id="@+id/text"
-        android:layout_gravity="start"
-        android:layout_marginStart="52dp"
-        android:textColor="#625899"
-        android:fontFamily="sans-serif-black"
-        android:textSize="40dp"
-        android:layout_marginLeft="52dp"
-        android:transitionName="welcome_text"/>
-
-    <EditText
-        android:layout_width="310dp"
-        android:layout_height="50dp"
-        android:background="@drawable/button"
-        android:id="@+id/edittext"
-        android:transitionName="login_text"
-        android:textColorHint="#625899"
-        android:hint="Email"
-        android:paddingLeft="20dp"
-        android:textSize="18dp"
-        android:fontFamily="sans-serif-medium"
-        android:layout_marginTop="20dp"/>
-
-    <EditText
-        android:layout_width="310dp"
-        android:layout_height="50dp"
-        android:background="@drawable/button"
-        android:transitionName="login_text"
-        android:id="@+id/edittext2"
-        android:hint="Password"
-        android:paddingLeft="20dp"
-        android:textColorHint="#625899"
-        android:textSize="18dp"
-        android:fontFamily="sans-serif-medium"
-        android:layout_marginTop="20dp"/>
-
-    <TextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="Forget Password?"
-        android:textSize="14dp"
-        android:transitionName="login_text2"
-        android:id="@+id/text2"
-        android:textColor="#625899"
-        android:layout_gravity="end"
-        android:paddingRight="55dp"
-        android:fontFamily="sans-serif"
-        android:layout_marginTop="10dp"/>
-
-    <Button
-        android:layout_width="310dp"
-        android:layout_height="50dp"
-        android:id="@+id/login_button2"
-        android:background="@drawable/filled"
-        app:backgroundTint="#652899"
-        android:transitionName="login_button"
-        android:text="@string/login"
-        android:textSize="18sp"
-        android:textColor="#ffffff"
-        android:fontFamily="sans-serif-medium"
-        android:layout_marginTop="20dp"/>
-
-    <TextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:id="@+id/text3"
-        android:text="@string/don_t_have_an_account_register_here"
-        android:textSize="14sp"
-        android:transitionName="login_text2"
-        android:textColor="#625899"
-        android:fontFamily="sans-serif"
-        android:layout_gravity="center"
-        android:layout_marginTop="15dp"/>
-
-</LinearLayout>
-
